+++ test/unit/Main.hs
@@ -3,7 +3,11 @@
 -- @since 0.1.0.0
 module Main (main) where
 
+import Data.List.NonEmpty (NonEmpty ((:|)))
+import Data.Version.Package (PackageVersion (MkPackageVersion, unPackageVersion))
+import Test.Tasty (TestTree, testGroup)
 import Test.Tasty qualified as Tasty
+import Test.Tasty.HUnit (testCase, (@=?))
 import Tests.IO qualified
 import Tests.TH qualified
 
@@ -16,5 +20,20 @@ main = do
     Tasty.testGroup
       "Unit tests"
       [ Tests.TH.tests,
-        Tests.IO.tests
+        Tests.IO.tests,
+        miscTests
       ]
+
+miscTests :: TestTree
+miscTests =
+  testGroup
+    "Misc tests"
+    [ testOverloadedRecordDot
+    ]
+
+testOverloadedRecordDot :: TestTree
+testOverloadedRecordDot = testCase "Compatible with OverloadedRecordDot" $ do
+  versNE @=? pv.unPackageVersion
+  where
+    versNE = 0 :| [1, 3]
+    pv = MkPackageVersion versNE